change in sctructure
